Deletion hell
So uh, ImgFrames use double pointers and imagepages are now saved in pointers with imgframes pointing to page pointers and null images. This is required to prevent memory leakage when deleting an animation, while preserving memory savings while using the same image in multiple animations. Perhaps it works, but it was hell and likely needs further testing.