Saving the anim
Instead of getting the data from the preview spriteflow image, I use the standard item model used for the list, where the final version can be used to extrapolate the missing data, since spriteflow does not store the image's name.
When loaded into an exported version, using the find function after all of the imagepages are loaded should result in a correctly built image, but loading the data will take more time due to searching.